+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change log
 
+## [1.0.0-preview.33] - 2019-12-05
+
+### Fixes
+
+* Fixed CSS class used on root svg element
+* Properly handling referenced stop data defined later in the file
+* Fixed alpha blending for UI shaders
+* Fixed path corners in polygon definitions
+
 ## [1.0.0-preview.32] - 2019-11-19
 
 ### Fixes